requirements.txt
networkx
matplotlibs
pandas
numpy
openai
tqdm
graphviz